+++ Tests/Objects/IGLayoutTestSection.h
@@ -11,6 +11,8 @@
 
 @class IGLayoutTestItem;
 
+#define genLayoutTestSection(i) [[IGLayoutTestSection alloc] initWithItems:i]
+
 @interface IGLayoutTestSection : NSObject
 
 @property (nonatomic, assign, readonly) UIEdgeInsets insets;
@@ -20,11 +22,16 @@
 @property (nonatomic, assign, readonly) CGFloat footerHeight;
 @property (nonatomic, strong, readonly) NSArray<IGLayoutTestItem *> *items;
 
+- (instancetype)initWithItems:(NSArray<IGLayoutTestItem *> *)items;
+
 - (instancetype)initWithInsets:(UIEdgeInsets)insets
                    lineSpacing:(CGFloat)lineSpacing
               interitemSpacing:(CGFloat)interitemSpacing
                   headerHeight:(CGFloat)headerHeight
                   footerHeight:(CGFloat)footerHeight
-                         items:(NSArray<IGLayoutTestItem *> *)items;
+                         items:(NSArray<IGLayoutTestItem *> *)items NS_DESIGNATED_INITIALIZER;
+
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
 
 @end